# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Financiamento Veículos",
        "description": "API para gerenciamento do sistema de financiamento de veículos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Lista todos os usuários",
                "description": "Retorna todos os usuários cadastrados",
                "operationId": "listarUsuarios",
                "responses": {
                    "200": {
                        "description": "Lista retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Cadastrar usuário",
                "operationId": "criarUsuario",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Edson Belém"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "edson@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "ativo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Buscar usuário",
                "operationId": "buscarUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID do usuário",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar usuário",
                "description": "Atualiza os dados cadastrais de um usuário",
                "operationId": "atualizarUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID do usuário",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Edson Belém"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "edson@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "ativo": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Excluir usuário",
                "description": "Remove um usuário do sistema",
                "operationId": "excluirUsuario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID do usuário",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário removido com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        },
        "/api/veiculos": {
            "get": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Lista todos os veículos",
                "operationId": "listarVeiculos",
                "responses": {
                    "200": {
                        "description": "Lista retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Cadastrar veículo",
                "operationId": "criarVeiculo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "placa",
                                    "marca",
                                    "modelo",
                                    "ano_fabricacao",
                                    "ano_modelo"
                                ],
                                "properties": {
                                    "placa": {
                                        "type": "string",
                                        "example": "ABC1D23"
                                    },
                                    "marca": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "modelo": {
                                        "type": "string",
                                        "example": "Corolla"
                                    },
                                    "ano_fabricacao": {
                                        "type": "integer",
                                        "example": 2024
                                    },
                                    "ano_modelo": {
                                        "type": "integer",
                                        "example": 2025
                                    },
                                    "cor": {
                                        "type": "string",
                                        "example": "Prata"
                                    },
                                    "renavam": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "chassi": {
                                        "type": "string",
                                        "example": "9BWZZZ377VT004251"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Veículo criado com sucesso"
                    }
                }
            }
        },
        "/api/veiculos/{id}": {
            "get": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Buscar veículo por ID",
                "operationId": "buscarVeiculo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veículo encontrado"
                    },
                    "404": {
                        "description": "Veículo não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Atualizar veículo",
                "operationId": "atualizarVeiculo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veículo atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Veículos"
                ],
                "summary": "Excluir veículo",
                "operationId": "excluirVeiculo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do veículo",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Veículo removido com sucesso"
                    }
                }
            }
        }
    },
...
}